Implement response DTO for GET /localdatetime endpoint
