+++ src/test/java/teammates/test/cases/webapi/GetLocalDateTimeInfoActionTest.java
@@ -7,6 +7,7 @@ import teammates.common.util.Const;
 import teammates.common.util.TimeHelper;
 import teammates.ui.webapi.action.GetLocalDateTimeInfoAction;
 import teammates.ui.webapi.action.JsonResult;
+import teammates.ui.webapi.output.LocalDateTimeInfoData;
 
 /**
  * SUT: {@link GetLocalDateTimeInfoAction}.
@@ -51,8 +52,7 @@ public class GetLocalDateTimeInfoActionTest extends BaseActionTest<GetLocalDateT
 
         GetLocalDateTimeInfoAction a = getAction(params);
         JsonResult r = getJsonResult(a);
-        GetLocalDateTimeInfoAction.LocalDateTimeInfo localDateTimeInfo =
-                (GetLocalDateTimeInfoAction.LocalDateTimeInfo) r.getOutput();
+        LocalDateTimeInfoData localDateTimeInfo = (LocalDateTimeInfoData) r.getOutput();
 
         assertEquals(TimeHelper.LocalDateTimeAmbiguityStatus.UNAMBIGUOUS, localDateTimeInfo.getResolvedStatus());
         assertEquals(1543669200000L, localDateTimeInfo.getResolvedTimestamp());
@@ -71,7 +71,7 @@ public class GetLocalDateTimeInfoActionTest extends BaseActionTest<GetLocalDateT
 
         a = getAction(params);
         r = getJsonResult(a);
-        localDateTimeInfo = (GetLocalDateTimeInfoAction.LocalDateTimeInfo) r.getOutput();
+        localDateTimeInfo = (LocalDateTimeInfoData) r.getOutput();
 
         assertEquals(TimeHelper.LocalDateTimeAmbiguityStatus.GAP, localDateTimeInfo.getResolvedStatus());
         assertEquals(1332639000000L, localDateTimeInfo.getResolvedTimestamp());
@@ -90,7 +90,7 @@ public class GetLocalDateTimeInfoActionTest extends BaseActionTest<GetLocalDateT
 
         a = getAction(params);
         r = getJsonResult(a);
-        localDateTimeInfo = (GetLocalDateTimeInfoAction.LocalDateTimeInfo) r.getOutput();
+        localDateTimeInfo = (LocalDateTimeInfoData) r.getOutput();
 
         assertEquals(TimeHelper.LocalDateTimeAmbiguityStatus.OVERLAP, localDateTimeInfo.getResolvedStatus());
         assertEquals(1351384200000L, localDateTimeInfo.getResolvedTimestamp());
